+++ src/main/java/org/encore/apartment/community/domain/apartment/data/dto/ResponseApartmentDto.java
@@ -9,21 +9,21 @@ import lombok.Data;
 @Data
 public class ResponseApartmentDto {
 
-    @NotNull
-    private Integer apartmentId;
+	@NotNull
+	private Long apartmentId;
 
-    @NotBlank
-    private String apartmentName;
+	@NotBlank
+	private String apartmentName;
 
-    @NotBlank
-    private String apartmentAddress;
-    private Integer apartmentTotalHousehold;
+	@NotBlank
+	private String apartmentAddress;
+	private Integer apartmentTotalHousehold;
 
-    public ResponseApartmentDto(Apartment params) {
-        this.apartmentId = params.getApartmentId();
-        this.apartmentName = params.getApartmentName();
-        this.apartmentAddress = params.getApartmentAddress();
-        this.apartmentTotalHousehold = params.getApartmentTotalHousehold();
-    }
+	public ResponseApartmentDto(Apartment params) {
+		this.apartmentId = params.getApartmentId();
+		this.apartmentName = params.getApartmentName();
+		this.apartmentAddress = params.getApartmentAddress();
+		this.apartmentTotalHousehold = params.getApartmentTotalHousehold();
+	}
 
 }
